requirements.txt
pydantic
PyMuPDF
python-docx
unstructured
tqdm
langchain
langdetect
pypika==0.48.9
openai==0.27.0
python-dotenv==0.21.0
llama-index==0.5.6
requests==2.28.2
pinecone==6.0.2
pinecone-client==2.0.0
msal
python-docx
python-pptx
pdfplumber
sentence-transformers